% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_visgraph}
\alias{plot_visgraph}
\title{Plot LASAGNA graph as interactive network with visNetwork}
\usage{
plot_visgraph(
  graph,
  layers = NULL,
  ntop = 100,
  min_rho = 0.3,
  mst = FALSE,
  vcex = 1,
  ecex = 1,
  egamma = 1,
  color.var = "value",
  edge.colors = c("blue", "purple"),
  labcex = 1,
  layout = NULL,
  physics = TRUE
)
}
\arguments{
\item{graph}{An igraph object.}

\item{layers}{Character vector of layers to include.}

\item{ntop}{Number of top nodes by absolute value.}

\item{min_rho}{Minimum absolute edge weight.}

\item{mst}{Use minimum spanning tree layout.}

\item{vcex}{Vertex size multiplier.}

\item{ecex}{Edge width multiplier.}

\item{physics}{Enable physics simulation.}
}
\value{
A visNetwork widget.
}
\description{
Plot LASAGNA graph as interactive network with visNetwork
}
